==> 02_Data_Concat.ipynb <==
# Data concatenating script

Authors: [Andreas Sørensen](https://www.linkedin.com/in/a-soerensen) and [Martin Röck](https://www.linkedin.com/in/martinroeck/) - Source: https://doi.org/10.5281/zenodo.5895051

# SETUP
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import time

print("Concatenating pre-processed data from DCT, CSTB and CLF...")

# Import data

Import data from csv to pandas dataframes and store them in one list

# IMPORT DATA FROM PRE-PROCESSED CSV FILES

#List of csv files to be imported
filename_list = ['DCT_processed.csv','CSTB_processed.csv','CLF_processed.csv']

#Initiation of dataframe list
df_list = []

#Load csv from the data processed folder
for filename in filename_list:
    df = pd.read_csv('00_data/1_data_pre_processed/'+filename, delimiter = ',')
    df_list.append(df)

# Concatenate data and export to CSV for feature engineering

# CONCATENATE DATAFRAMES

#Concatenate all dataframes in df_list in the order they are loaded
df_concat = df_list[0].append(df_list[1:], ignore_index=True, sort=False)

#Replace NaN with 'n/a' for completeness
df_concat.fillna('n/a',inplace=True)

#Create csv
filename = '00_data/2_data_concatenated/EU-ECB_dataset_concatenated.csv'
df_concat.to_csv(filename, index=False)

print("Data from DCT, CSTB and CLF concatenated.")
